+++ src/features/user/components/create-user-form.tsx
@@ -6,7 +6,7 @@ import { DatePicker } from '@/components/date-picker';
 import { ToggleGroup } from '@/components/toggle-group';
 import { Button } from '@/components/button';
 import { ButtonGroup } from '@/components/button-group';
-import { Role, roleOptions } from '@/lib/auth';
+import { Role, roleOptions } from '@/lib/roles';
 import { isInPast } from '@/lib/dates';
 
 const schema = z
